+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 - added poppins font in Styleguide !2564 @Niklas003 
 - Use email address DTO and let the MailboxGateway handle JSON encoding !2520 @alex.simm
 - Replaced the usage of a function in FoodsaverGateway with a more ressource saving function !2595 @alex.simm
-- Use Store DTO in StoreGateway !2515 @KoeMai
+- Use Store DTO in StoreGateway !2515 !2611 @KoeMai
 - Move business logic for editing store into StoreTransaction !2515 @KoeMai
 - Fix TeamStoreStatus and move getStoreMarkers to StoreGateway !2560 @KoeMai